refactor: 테이블과과 관련된 클래스들을 table 패키지로 이동
