+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.table.domain;
 
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
