+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -17,7 +17,7 @@ import kitchenpos.menu.fixture.MenuProductFixture;
 import kitchenpos.application.fixture.OrderFixture;
 import kitchenpos.application.fixture.OrderLineItemFixture;
 import kitchenpos.application.fixture.ProductFixture;
-import kitchenpos.application.fixture.TableFixture;
+import kitchenpos.table.fixture.TableFixture;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
@@ -27,8 +27,8 @@ import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderLineItemRepository;
 import kitchenpos.domain.OrderRepository;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.Quantity;
 import org.junit.jupiter.api.BeforeEach;
