+++ src/test/java/kitchenpos/table/fixture/TableFixture.java
@@ -1,7 +1,7 @@
-package kitchenpos.application.fixture;
+package kitchenpos.table.fixture;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 
 public class TableFixture {
